auth and cli added

+++ controllers/cards_controller.py
@@ -1,12 +1,13 @@
 from flask import Blueprint, jsonify, request
 from datetime import date
-from db import db
+from init import db
 from models.card import Card, CardSchema
+from flask_jwt_extended import jwt_required
+from controllers.auth_controller import authorize
 
 cards_bp = Blueprint('cards', __name__, url_prefix='/cards')
 
 @cards_bp.route('/')
-# @jwt_required()
 def get_all_cards():
     # if not authorize():
     #     return {'error': 'You must be an admin'}, 401
@@ -25,6 +26,7 @@ def get_one_card(id):
         return {'error': f'Card not found with id {id}'}, 404
 
 @cards_bp.route('/', methods=['POST'])
+@jwt_required()
 def create_card():
     # Create a new card model instance from the user_info
     card = Card(
@@ -41,9 +43,11 @@ def create_card():
     return CardSchema().dump(card), 201
 
 @cards_bp.route('/<int:id>/', methods=['PUT', 'PATCH'])
+@jwt_required()
 def update_one_card(id):
     stmt = db.select(Card).filter_by(id=id)
     card = db.session.scalar(stmt)
+
     if card:
         card.title = request.json.get('title') or card.title
         card.description = request.json.get('description') or card.description
@@ -55,9 +59,13 @@ def update_one_card(id):
         return {'error': f'Card not found with id {id}'}, 404
 
 @cards_bp.route('/<int:id>/', methods=['DELETE'])
+@jwt_required()
 def delete_one_card(id):
+    authorize()
+
     stmt = db.select(Card).filter_by(id=id)
     card = db.session.scalar(stmt)
+
     if card:
         db.session.delete(card)
         db.session.commit()
